tests: add page tests
